requirements.txt
diffusers==0.26.2
huggingface_hub==0.20.3
numpy==1.26.4
Pillow==9.3.0
skimage==0.0
streamlit==1.18.1
torch==2.0.1+cu118
torchvision==0.15.2
